Move GooglePayRepository to googlepaylauncher package
This is a more appropriate package for `GooglePayLauncher`
and `DefaultGooglePayLauncher`
