+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/GooglePayPaymentMethodLauncherViewModel.kt
@@ -18,8 +18,6 @@ import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.networking.ApiRequest
 import com.stripe.android.networking.StripeApiRepository
 import com.stripe.android.networking.StripeRepository
-import com.stripe.android.paymentsheet.DefaultGooglePayRepository
-import com.stripe.android.paymentsheet.GooglePayRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import org.json.JSONObject
