--- payments-core/src/main/java/com/stripe/android/paymentsheet/GooglePayRepository.kt
@@ -1,12 +1,0 @@
-package com.stripe.android.paymentsheet
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
-
-fun interface GooglePayRepository {
-    fun isReady(): Flow<Boolean>
-
-    object Disabled : GooglePayRepository {
-        override fun isReady(): Flow<Boolean> = flowOf(false)
-    }
-}
