+++ payments-core/src/main/java/com/stripe/android/paymentsheet/injection/PaymentSheetViewModelModule.kt
@@ -4,13 +4,15 @@ import android.app.Application
 import android.content.Context
 import com.stripe.android.Logger
 import com.stripe.android.PaymentConfiguration
+import com.stripe.android.googlepaylauncher.DefaultGooglePayRepository
+import com.stripe.android.googlepaylauncher.GooglePayEnvironment
+import com.stripe.android.googlepaylauncher.GooglePayRepository
 import com.stripe.android.networking.ApiRequest
 import com.stripe.android.networking.StripeApiRepository
 import com.stripe.android.payments.core.injection.ENABLE_LOGGING
 import com.stripe.android.payments.core.injection.IOContext
-import com.stripe.android.paymentsheet.DefaultGooglePayRepository
 import com.stripe.android.paymentsheet.DefaultPrefsRepository
-import com.stripe.android.paymentsheet.GooglePayRepository
+import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PaymentSheetContract
 import com.stripe.android.paymentsheet.PrefsRepository
 import com.stripe.android.paymentsheet.model.ClientSecret
@@ -100,7 +102,12 @@ internal abstract class PaymentSheetViewModelModule {
             return starterArgs.config?.googlePay?.environment?.let { environment ->
                 DefaultGooglePayRepository(
                     appContext,
-                    environment
+                    when (environment) {
+                        PaymentSheet.GooglePayConfiguration.Environment.Production ->
+                            GooglePayEnvironment.Production
+                        PaymentSheet.GooglePayConfiguration.Environment.Test ->
+                            GooglePayEnvironment.Test
+                    }
                 )
             } ?: GooglePayRepository.Disabled
         }
