+++ payments-core/src/test/java/com/stripe/android/googlepaylauncher/GooglePayPaymentMethodLauncherTest.kt
@@ -13,10 +13,7 @@ import androidx.fragment.app.testing.launchFragmentInContainer
 import androidx.lifecycle.Lifecycle
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.model.PaymentMethodFixtures
-import com.stripe.android.paymentsheet.GooglePayRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.TestCoroutineScope
 import org.junit.runner.RunWith
@@ -110,12 +107,6 @@ class GooglePayPaymentMethodLauncherTest {
                 )
             }
         }
-    }
-
-    private class FakeGooglePayRepository(
-        private var isReady: Boolean
-    ) : GooglePayRepository {
-        override fun isReady(): Flow<Boolean> = flowOf(isReady)
     }
 
     private class FakeActivityResultRegistry(
