+++ payments-core/src/test/java/com/stripe/android/paymentsheet/PaymentSheetActivityTest.kt
@@ -17,6 +17,7 @@ import com.stripe.android.StripeIntentResult
 import com.stripe.android.StripePaymentController
 import com.stripe.android.databinding.PrimaryButtonBinding
 import com.stripe.android.databinding.StripeGooglePayButtonBinding
+import com.stripe.android.googlepaylauncher.FakeGooglePayRepository
 import com.stripe.android.googlepaylauncher.GooglePayLauncherResult
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.ConfirmPaymentIntentParams
